+++ pyxel/detectors/geometry.py
@@ -8,6 +8,8 @@
 """Geometry class for detector."""
 import numpy as np
 
+from pyxel.util.memory import get_size
+
 
 class Geometry:
     """Geometrical attributes of the detector."""
@@ -55,6 +57,8 @@ class Geometry:
         self._total_thickness = total_thickness
         self._pixel_vert_size = pixel_vert_size
         self._pixel_horz_size = pixel_horz_size
+
+        self._numbytes = 0
 
     def __repr__(self) -> str:
         cls_name = self.__class__.__name__  # type: str
@@ -167,3 +171,15 @@ class Geometry:
         init_hor_position = np.arange(0.0, self.col, 1.0) * self.pixel_horz_size
         init_hor_position += self.pixel_horz_size / 2.0
         return np.tile(init_hor_position, self.row)
+
+    @property
+    def numbytes(self) -> int:
+        """Recursively calculates object size in bytes using Pympler library.
+
+        Returns
+        -------
+        int
+            Size of the object in bytes.
+        """
+        self._numbytes = get_size(self)
+        return self._numbytes