+++ pyxel/detectors/material.py
@@ -13,6 +13,8 @@ from pathlib import Path
 
 import numpy as np
 
+from pyxel.util.memory import get_size
+
 
 class MaterialType(Enum):
     """TBW."""
@@ -109,6 +111,7 @@ class Material:
         self._e_effective_mass = (
             e_effective_mass  # TODO: set automatically depending on the material
         )
+        self._numbytes = 0
 
     @property
     def trapped_charge(self) -> t.Optional[np.ndarray]:
@@ -207,6 +210,18 @@ class Material:
             raise ValueError("'e_effective_mass' must be between 0.0 and 1.e-10.")
 
         self._e_effective_mass = value
+
+    @property
+    def numbytes(self) -> int:
+        """Recursively calculates object size in bytes using Pympler library.
+
+        Returns
+        -------
+        int
+            Size of the object in bytes.
+        """
+        self._numbytes = get_size(self)
+        return self._numbytes
 
     # TODO create func for compound materials
     # def set_material(self, material):