+++ CHANGELOG.rst
@@ -28,6 +28,8 @@ Core
   (See `!132 <https://gitlab.com/esa/pyxel/-/merge_requests/132>`_).
 * Reimplement convolution in POPPY optical_psf model.
   (See `#52 <https://gitlab.com/esa/pyxel/issues/52>`_).
+* Add property 'Detector.numbytes' and/or method 'Detector.memory_usage()'
+  (See `!116 <https://gitlab.com/esa/pyxel/-/merge_requests/116>`_).
 
 Documentation
 ~~~~~~~~~~~~~